WebSocket Configuration.
