--- src/main/java/com/voicecaster/voicecaster/Greeting.java
@@ -1,16 +1,0 @@
-package com.voicecaster.voicecaster;
-
-public class Greeting {
-    private String content;
-    public Greeting() {
-
-    } 
-
-    public Greeting(String content) {
-        this.content = content;
-    }
-
-    public String getContent() {
-        return content;
-    }
-}
