+++ src/main/java/com/voicecaster/voicecaster/models/OutputMessage.java
@@ -1,0 +1,17 @@
+package com.voicecaster.voicecaster.models;
+
+public class OutputMessage {
+    private String content;
+
+    public OutputMessage() {
+
+    }
+
+    public OutputMessage(String content) {
+        this.content = content;
+    }
+
+    public String getContent() {
+        return content;
+    }
+}
