+++ src/main/java/com/voicecaster/voicecaster/wsconfig/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package com.voicecaster.voicecaster;
+package com.voicecaster.voicecaster.wsconfig;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
@@ -8,8 +8,8 @@ import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerCo
 
 @Configuration
 @EnableWebSocketMessageBroker
-public class WebSocketConfig implements WebSocketMessageBrokerConfigurer{
-
+public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
+    
     @Override
     public void configureMessageBroker(MessageBrokerRegistry config) {
         config.enableSimpleBroker("/topic");
@@ -18,7 +18,6 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer{
 
     @Override
     public void registerStompEndpoints(StompEndpointRegistry registry) {
-        registry.addEndpoint("/gs-guide-websocket").setAllowedOrigins("http://localhost:5000").withSockJS();
+        registry.addEndpoint("/ws").setAllowedOrigins("http://192.168.1.244:8080", "chrome-extension://kmiiilhamlibadehdjnpdlhlppplkmfm").withSockJS();
     }
-    
 }
