+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-server.port = 8090
+server.port = 9000
